Add additional validation

+++ src/auth.js
@@ -1,24 +1,56 @@
 import { auth, firestore, storage } from './firebase';
 
+const avatarFileTypes = [
+  'image/gif',
+  'image/jpeg',
+  'image/png',
+  'image/webp',
+  'image/svg+xml'
+];
+
 export const changeAvatar = (avatar) => {
   return new Promise((resolve, reject) => {
     if (!avatar) {
       reject();
+
+      return;
+    }
+
+    if (!avatarFileTypes.includes(avatar.type)) {
+      reject();
+
+      return;
+    }
+
+    if (avatar.size > (20 * 1024 * 1024)) {
+      reject();
+
+      return;
     }
   
     const currentUser = auth.currentUser;
   
     if (!currentUser) {
       reject();
+
+      return;
     }
   
     const uid = currentUser.uid;
   
     if (!uid) {
       reject();
+
+      return;
     }
 
     const reference = storage.ref().child('images').child('avatars').child(uid);
+
+    if (!reference) {
+      reject();
+
+      return;
+    }
 
     reference.put(avatar).then((uploadTaskSnapshot) => {
       reference.getDownloadURL().then((value) => {
@@ -44,18 +76,28 @@ export const removeAvatar = () => {
 
     if (!currentUser) {
       reject();
+
+      return;
     }
 
     const uid = currentUser.uid;
 
     if (!uid) {
       reject();
+
+      return;
     }
 
     currentUser.updateProfile({
       photoURL: null
     }).then((value) => {
       const reference = storage.ref().child('images').child('avatars').child(uid);
+
+      if (!reference) {
+        reject();
+
+        return;
+      }
 
       reference.delete().then((value) => {
         resolve(value);
@@ -72,18 +114,24 @@ export const changeFirstName = (firstName) => {
   return new Promise((resolve, reject) => {
     if (!firstName) {
       reject();
+
+      return;
     }
   
     const currentUser = auth.currentUser;
   
     if (!currentUser) {
       reject();
+
+      return;
     }
   
     const uid = currentUser.uid;
   
     if (!uid) {
       reject();
+
+      return;
     }
 
     firestore.collection('users').doc(uid).update({
@@ -100,18 +148,24 @@ export const changeLastName = (lastName) => {
   return new Promise((resolve, reject) => {
     if (!lastName) {
       reject();
+
+      return;
     }
   
     const currentUser = auth.currentUser;
   
     if (!currentUser) {
       reject();
+      
+      return;
     }
   
     const uid = currentUser.uid;
   
     if (!uid) {
       reject();
+
+      return;
     }
 
     firestore.collection('users').doc(uid).update({
@@ -128,18 +182,24 @@ export const changeUsername = (username) => {
   return new Promise((resolve, reject) => {
     if (!username) {
       reject();
+
+      return;
     }
   
     const currentUser = auth.currentUser;
   
     if (!currentUser) {
       reject();
+
+      return;
     }
   
     const uid = currentUser.uid;
   
     if (!uid) {
       reject();
+
+      return;
     }
 
     firestore.collection('users').doc(uid).update({
@@ -156,18 +216,24 @@ export const changeEmailAddress = (emailAddress) => {
   return new Promise((resolve, reject) => {
     if (!emailAddress) {
       reject();
+
+      return;
     }
   
     const currentUser = auth.currentUser;
   
     if (!currentUser) {
       reject();
+
+      return;
     }
   
     const uid = currentUser.uid;
   
     if (!uid) {
       reject();
+
+      return;
     }
 
     currentUser.updateEmail(emailAddress).then((value) => {
@@ -184,6 +250,8 @@ export const verifyEmailAddress = () => {
 
     if (!currentUser) {
       reject();
+
+      return;
     }
 
     currentUser.sendEmailVerification().then((value) => {
